Webhook Support + Mention Fixes

+++ src/raid_coordinator/bot.py
@@ -1,14 +1,9 @@
 import asyncio
 import discord
-import functools
-import json
 import re
-import sys
-import time
 import traceback
 import pytz
-from collections import namedtuple
-from datetime import datetime, timedelta, timezone
+from datetime import datetime, timedelta
 
 
 # buffer for busier servers
@@ -63,19 +58,19 @@ def get_active_raids_channel(server):
     return discord.utils.find(lambda c: c.name == settings.active_raids_channel_name, server.channels)
 
 
-def encode_message(message):
+def encode_message(creator, message):
     """Encodes a message for storage in the channel topic."""
-    return "{}|{}".format(message.channel.id, message.id)
+    return "{}|{}|{}".format(creator.id, message.channel.id, message.id)
 
 
 def decode_message(data):
     """Decodes a channel topic into a channel id, message id tuple."""
     try:
-        channel_id, message_id = data.split('|')
+        creator_id, channel_id, message_id = data.split('|')
     except:
-        return None, None
+        return None, None, None
 
-    return channel_id, message_id
+    return creator_id, channel_id, message_id
 
 
 def lookup_raid_channel(message):
@@ -90,7 +85,7 @@ def lookup_raid_channel(message):
     server = message.server
     if message.channel_mentions:
         raid_channel = message.channel_mentions[0]
-        channel_id, message_id = decode_message(raid_channel.topic)
+        _, channel_id, message_id = decode_message(raid_channel.topic)
         if channel_id and message_id:
             return raid_channel
 
@@ -156,6 +151,11 @@ def get_raid_role(channel):
     return discord.utils.find(lambda r: r.name == channel.name, channel.server.roles)
 
 
+def created_by_bot(channel):
+    creator = get_raid_creator(channel)
+    return creator is None or creator.bot
+
+
 def get_raid_members(channel):
     """Gets the raid members in this channel."""
     return [target for target, _ in channel.overwrites if isinstance(target, discord.User)]
@@ -166,7 +166,7 @@ def get_raid_start_embed(creator, started_dt, expiration_dt):
     embed = discord.Embed()
     embed.color = discord.Color.green()
     embed.title = 'A raid has started!'
-    embed.add_field(name='creator', value=creator.mention, inline=True)
+    embed.add_field(name='creator', value=creator.display_name, inline=True)
     embed.add_field(name='started at', value=started_dt.strftime(settings.time_format), inline=False)
     embed.add_field(name='channel expires', value=expiration_dt.strftime(settings.time_format), inline=False)
     embed.set_footer(text='To join, tap {} below'.format(get_join_emoji()))
@@ -184,13 +184,13 @@ def get_raid_active_embed(num_members, started_dt, expiration_dt):
     return embed
 
 
-def get_raid_end_embed(creator, started_dt, ended_dt):
+def get_raid_end_embed(creator, started_dt, ended_dt, default_creator=None):
     """Constructs an embed for the end of a raid."""
     duration = ended_dt - started_dt
     embed = discord.Embed()
     embed.color = discord.Color.red()
     embed.title = 'This raid has ended.'
-    embed.add_field(name='creator', value=creator.mention if creator else None, inline=True)
+    embed.add_field(name='creator', value=creator.display_name if creator else default_creator, inline=True)
     embed.add_field(name='duration', value=strfdelta(duration, '{hours:02}:{minutes:02}:{seconds:02}'), inline=True)
     embed.add_field(name='started at', value=started_dt.strftime(settings.time_format), inline=False)
     embed.add_field(name='ended at', value=ended_dt.strftime(settings.time_format), inline=False)
@@ -236,7 +236,7 @@ def get_raid_summary_embed(creator, channel_name, expiration_dt, text):
     embed = discord.Embed()
     embed.title = 'Welcome to this raid channel!'
     embed.description = "**{}**".format(text)
-    embed.add_field(name='creator', value=creator.mention)
+    embed.add_field(name='creator', value=creator.display_name)
     embed.add_field(name='channel expires', value=expiration_dt.strftime(settings.time_format))
     embed.add_field(name='raid group', value='@{}'.format(channel_name), inline=False)
     embed.add_field(name="commands", value="You can use the following commands:", inline=False)
@@ -272,7 +272,7 @@ def is_open(channel):
 async def get_announcement_message(raid_channel):
     """Gets the message that created this channel."""
     server = raid_channel.server
-    channel_id, message_id = decode_message(raid_channel.topic)
+    _, channel_id, message_id = decode_message(raid_channel.topic)
     try:
         channel = server.get_channel(channel_id)
         if channel:
@@ -282,27 +282,11 @@ async def get_announcement_message(raid_channel):
         return None  # an error occurred, return None TODO: log here
 
 
-async def get_raid_creator(raid_channel):
-    """Gets the user who created the raid.
-
-    Pulls and resolves this information from the announcement embed.
+def get_raid_creator(raid_channel):
+    """Gets the member who created the raid.
     """
-    message = await get_announcement_message(raid_channel)
-    if message is not None and message.embeds:
-        embed = message.embeds[0]
-        fields = embed.get('fields', [])
-        if fields:
-            creator_mention = fields[0]['value']
-
-            # try in the overwrites
-            for target, _ in raid_channel.overwrites:
-                if isinstance(target, discord.User) and target.mention == creator_mention:
-                    return target
-
-            # otherwise try in the server users (less efficient but right)
-            for target in raid_channel.server.members:
-                if isinstance(target, discord.User) and target.mention == creator_mention:
-                    return target
+    creator_id, _, _ = decode_message(raid_channel.topic)
+    return raid_channel.server.get_member(creator_id)
 
 
 def is_expired(message):
@@ -332,8 +316,8 @@ async def start_raid_group(user, message, description):
     """Starts a new raid group."""
     global should_refresh_active_raids
 
-    # get the server
-    server = user.server
+    # get the server, use the message because user might be a webhook with no server
+    server = message.server
 
     # find an available raid channel
     channel = get_available_raid_channel(server)
@@ -347,7 +331,7 @@ async def start_raid_group(user, message, description):
 
         try:
             # set the topic
-            await client.edit_channel(channel, topic=encode_message(message))
+            await client.edit_channel(channel, topic=encode_message(user, message))
 
             # create a role with the same name as this channel
             role = await client.create_role(server, name=channel.name, mentionable=True)
@@ -375,6 +359,18 @@ async def start_raid_group(user, message, description):
         return channel
 
 
+async def get_original_creator_name(raid_channel):
+    """Gets the user who created the raid.
+    Pulls and resolves this information from the announcement embed.
+    """
+    message = await get_announcement_message(raid_channel)
+    if message is not None and message.embeds:
+        embed = message.embeds[0]
+        fields = embed.get('fields', [])
+        if fields:
+            return fields[0]['value']
+
+
 async def end_raid_group(channel):
     """Ends a raid group."""
     global should_refresh_active_raids
@@ -382,7 +378,10 @@ async def end_raid_group(channel):
     server = channel.server
 
     # get the creator before we remove roles
-    creator = await get_raid_creator(channel)
+    creator = get_raid_creator(channel)
+    original_creator_name = None
+    if creator is None:
+        original_creator_name = await get_original_creator_name(channel)
 
     # remove all the permissions
     role = await get_raid_viewer_role(server)
@@ -406,7 +405,7 @@ async def end_raid_group(channel):
     if message is not None:
         started_dt = adjusted_datetime(message.timestamp)
         ended_dt = adjusted_datetime(datetime.utcnow())
-        await client.edit_message(message, embed=get_raid_end_embed(creator, started_dt, ended_dt))
+        await client.edit_message(message, embed=get_raid_end_embed(creator, started_dt, ended_dt, original_creator_name))
         await client.clear_reactions(message)
 
     # remove the topic
@@ -418,6 +417,10 @@ async def end_raid_group(channel):
 
 async def invite_user_to_raid(channel, user):
     """Invites a user to the raid channel."""
+    # don't invite bots and webhooks
+    if user.bot:
+        return
+
     # adds an overwrite for the user
     perms = discord.PermissionOverwrite(read_messages=True)
     await client.edit_channel_permissions(channel, user, perms)
@@ -430,14 +433,18 @@ async def invite_user_to_raid(channel, user):
     # sends a message to the raid channel the user was added
     await client.send_message(channel,
                               "{}, you are now a member of this raid group.".format(user.mention),
-                              embed=get_success_embed('{} has joined the raid!'.format(user.mention)))
+                              embed=get_success_embed('{} has joined the raid!'.format(user.display_name)))
 
 
 async def uninvite_user_from_raid(channel, user):
     """Removes a user from a raid channel."""
+    # skip bots and webhooks
+    if user.bot:
+        return
+
     # reflect the proper number of members (the bot role and everyone are excluded)
     await client.delete_channel_permissions(channel, user)
-    await client.send_message(channel, embed=get_error_embed('{} has the left raid!'.format(user.mention)))
+    await client.send_message(channel, embed=get_error_embed('{} has the left raid!'.format(user.display_name)))
 
     # delete user from role
     role = get_raid_role(channel)
@@ -510,7 +517,7 @@ async def cleanup_raid_channels():
                 for channel in channels:
                     if channel not in locked_channels and not is_open(channel):
                         message = await get_announcement_message(channel)
-                        if is_expired(message) or not get_raid_members(channel):
+                        if is_expired(message) or (not created_by_bot(channel) and not get_raid_members(channel)):
                             await end_raid_group(channel)
 
                 # list the active raids every cycle
@@ -647,7 +654,7 @@ async def on_message(message):
         if is_organizer:
             await end_raid_group(channel)
         else:
-            creator = await get_raid_creator(channel)
+            creator = get_raid_creator(channel)
             if user == creator:
                 await end_raid_group(channel)
             else:
